requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
opencv-python>=4.5.0
Pillow>=8.0.0
PyYAML>=6.0
pyinstaller>=5.0.0
xlsxwriter>=3.0.0
